requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
blinker==1.4
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831756778/work
chardet @ file:///C:/ci/chardet_1607706937985/work
click==8.0.3
colorama==0.4.4
conda==4.10.3
conda-package-handling @ file:///C:/ci/conda-package-handling_1618262410900/work
coverage==6.1.2
cryptography @ file:///C:/ci/cryptography_1616769504165/work
Django==3.2.9
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
Flask==2.0.2
Flask-Mail==0.9.1
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata==4.8.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
menuinst==1.4.16
mysql-connector-python==8.0.27
packaging==21.3
pluggy==1.0.0
protobuf==3.19.1
psycopg2==2.9.2
py==1.11.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyodbc==4.0.32
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==3.0.6
PySocks @ file:///C:/ci/pysocks_1605307512533/work
pytest==6.2.5
pytz==2021.3
pywin32==228
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
sqlparse==0.4.2
toml==0.10.2
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
Werkzeug==2.0.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
zipp==3.6.0
